♻️ refactor: Role 클래스 엔티티 외부로 뺌
