+++ src/main/java/com/project/cheerha/domain/user/entity/Role.java
@@ -1,0 +1,5 @@
+package com.project.cheerha.domain.user.entity;
+
+public enum Role {
+    USER, ADMIN
+}
